Upgrade psutil (5.9.8) for arm macOS wheels (#20760)
This upgrades Pants' dependencies psutil 5.9.0
-> 5.9.8 because the new version has pre-built wheels for ARM
platforms (ijson previously had ARM linux, but not macOS; psutil had
neither).

This papers over the cache weirdness in #20759 / #20765, which
is related to having to build `psutil` native code on ARM macOS, and
the builder building x86-64 code by default at the moment.

This is a partial cherry-pick of #20760, just psutil, to minimise the
impact/risk.