Upgrade ijson (3.2.3) and psutil (5.9.8) for arm macOS wheels (#20760)
This upgrades Pants' dependencies ijson 3.1.4 -> 3.2.3 and psutil 5.9.0
-> 5.9.8 because the new versions have pre-built wheels for ARM
platforms (ijson previously had ARM linux, but not macOS; psutil had
neither).

This might've reduced the chance of the cache weirdness in #20759, which
was related to `psutil` on ARM macOS, and so maybe was related to having
to build the wheels. Clearing the cache seemed to resolve the acute
issue, though. 🤷‍♂️

There's still two dependencies that don't have wheels (and they don't
have them for any platform):

- `python-multipart` (newer versions do, but
`strawberry-graphql==0.114.0` pins this to 0.0.5, so upgrading is likely
a bit more fiddle)
- `pydevd-pycharm` (the newest version still only provides an sdist, no
wheels)

Both of these appear to be pure Python (i.e. `pip install`'s automatic
build results in `...-py3-none-any.whl` files) and thus having pre-built
wheels is less valuable. The two upgraded in this PR do have native code
and thus have platform-dependent wheels.