Fix direct_url in python PEP660 editable wheels
We were using the first entry in `source_roots_result.path_to_root` to
populate the editable wheel's `direct_url`. However, that does not work
as I expected it to because `SourceRootsRequest.dirs` and
`SourceRootsResult.path_to_root` both get sorted as soon as the
dataclass is created (if not before).

PR #20486 was my first attempt at fixing this, but that wasn't enough.
Since the heuristics used before #20486 and in #20486 were both
inaccurate due to dataclasses getting sorted, this abandons the
heuristics altogether, explicitly requesting the source root for the
`python_distribution` target.