Improve exported mutable_virtualenv prompt (#20796)
This includes resolve name and python version in the PS1 prompt of
`mutable_virtualenv` exports. This does not make the prompt
configurable, it just improves the default.

The implementation relies on pex's `venv --prompt=` cli arg, which
passes it to its `virtualenv` (or `-m venv`) subprocess:
-
https://github.com/pex-tool/pex/blob/v2.3.1/pex/venv/installer_options.py#L86-L89
-
https://github.com/pex-tool/pex/blob/v2.3.1/pex/venv/virtualenv.py#L215
-
https://github.com/pex-tool/pex/blob/v2.3.1/pex/venv/virtualenv.py#L252

Related: #17727
(#17727 requests an improved prompt for
`symlinked_immutable_virtualenv`. This PR adds that for
`mutable_virtualenv` not for `symlinked_immutable_virtualenv`)