Fix direct_url in python PEP660 editable wheels (again) (#20798)
We were using the first entry in `source_roots_result.path_to_root` to
populate the editable wheel's `direct_url`. However, that does not work
as I expected it to because `SourceRootsRequest.dirs` and
`SourceRootsResult.path_to_root` both get sorted as soon as the
dataclass is created (if not before).

PR #20486 was my first attempt at fixing this, but that wasn't enough.
Since the heuristics used before #20486 and in #20486 were both
inaccurate due to dataclasses getting sorted, this abandons the
heuristics altogether, explicitly requesting the source root for the
`python_distribution` target.

This time, I added a test to make sure direct_url.json has the contents
we expect.

Note: This is ultimately a cosmetic bug. The direct_url gets printed in
the output of `pip list`, but is otherwise not really exposed anywhere.
So, it's odd when it's the wrong directory, but does not break any
functionality.