Introduce a FromfileExpander struct (#20811)
As with #20807, this is another preliminary change to support
fromfile paths being relative to the buildroot. 

This change introduces a FromfileExpander struct, and moves
the expansion logic from free functions to struct methods. 
The free functions are then reimplemented as calls into 
these methods, to avoid perturbing calling code.

A followup change will plumb instances of this struct
wherever needed, and remove the free functions.

Another followup change will introduce the buildroot
context to this struct, to finally achieve the underlying
functionality change.

This change is just a refactoring, to make the subsequent
changes easier to grok.