Remove deprecated `pex_binary(platforms=...)` for 2.22, deprecate `resolve_local_platforms` (#20823)
This field was deprecated in #20012, released in 2.19. We've now
branched for 2.21 and thus `main` is prepping for 2.22. So, the
deprecation has expired!

In that PR, there was discussion about working out whether this should
remain for longer. This has been deprecated for several releases now,
and it's being removed early in the 2.22 release cycle: if someone
comments, we can reinstate/reconsider.

---

Also, as part of this, I noticed the `resolve_local_platforms` field and
option now seem to be meaningless: the underlying `pex
--resolve-local-platforms` field is documented as `When --platforms are
specified, ...`. Thus, this deprecates that field and the global option
for 2.24.