Bundle OpenAPI files as a single resource (#20695)
An old branch I decided to dust off and finish.  🧹

The OpenAPI targets were a great addition to allow futher
OpenAPI-related functionality in Pants (codegen, linters, etc.), but I
failed to dogfood them for my own use case. My problem with the current
implementation is that it's hard to use an `openapi_document` as a
dependency of a `python_source` due to it behaving more or less like a
`file()` would.

At work we currently have all our API descriptions as `resource()` so
they're bundled with the Python code that uses them for request
validation and whatnot. With `openapi_document` this would require you
to manually specify your `openapi_document` and all its `openapi_source`
dependencies as inputs to `experimental_wrap_as_resources`, which is
annoying when the OpenAPI targets all have dependency inference.

This PR uses codegen and `redocly bundle` to bundle an
`openapi_document` and its `openapi_source`s into a single file when the
`openapi_document` is needed as a dependency by another target. This
allow developers to keep their multi-file approach when writing the API
description, existing OpenAPI backends will (should?) continue to work
as-is, but when consuming it elsewhere as a dependency they now only
have to care about one single OpenAPI file.

I initially wrote this code back when `experimental_wrap_as_resources`
didn't exist and `openapi_document` was (in our case) unusable, which is
why I made it part of the main OpenAPI backend (and thus enforced).
Nowadays you can, as I mentioned, likely get work around the issue with
`experimental_wrap_as_resources` - so maybe it's better to move this
functionality into its own, opt-in, backend? That being said, resolving
`$ref`s in a multi-file approach is _tricky_ to get right, which is why
a lot of OpenAPI tools don't even bother supporting it (or behave
differently), so Pants "always" passing a single file around could be
seen as a good thing. 🤷

Edit: `redocly bundle` could also be useful as a package goal to prepare
the API definition for distribution by dereferencing includes, stripping
`x-internal` paths etc., but that's a separate feature.