An option for excluding certain files from all deploy jars (#20943)
The default value will strip signature files from deploy jars, which is
necessary since the deploy jar will naturally not match the signature
of some random jar that was splatted in to it.

The alternative was to set this as the default on `deploy_jar`'s
`exclude_files` field. But then anyone setting that field would 
overrwrite the default, and be surprised when everything broke
due to their being *more* files in their jar than before.

The previous workaround was to set these patterns manually 
on each deploy_jar, but that is an unnecessary gotcha.