Mark most JVM-based subsystems as exportable (#20788)
Converts all JVM tools which are subsystems to use ExportableTool for
lockfile generation.

Some internal tools still use `GenerateJvmToolLockfileSentinel`
subclasses. They aren't subsystems, so they'd need some more work. We
could either convert them into subsystems (as is done in Python-based
backends) although if we don't plan on exposing them (which is the
current case) we could also remove the subclass and rule and replace it
with a function.