Expose intrinsics as decorated Python functions instead (#20874)
Remove `intrinsics` facility from the engine, and instead expose
intrinsics in `intrinsics.py` via a new native call response for
generators. This solves an issue in #19730 that there were no obvious
`@rule` functions to call for intrinsics.

The native call response wraps a Rust future, which we trampoline to
execute on the Rust runtime, and then return a value to the generator.
It should work generically for any use case where we would like `@rule`
code to await native code, which opens the door to fixing #11329 by
making
[`Workspace.write_digest`](https://github.com/pantsbuild/pants/blob/00c757e77f3509dec2849982b148d41d95d6e629/src/python/pants/engine/fs.py#L290-L308)
async.