export generated python sources to exported venvs (#20975)
Issue: A user has developers who use VSCode and would like those
developers to have easy access to the generated Python sources for
`protobuf_sources` targets in VSCode. The user reported that their
developers found having to run `pants export` and also `pants
export-codegen` confusing and unwiedly. The user would prefer a simpler
DX involving just using `pants export`.

When exporting a mutable venv for a resolve, write generated Python
codegen products to the venv as well when
`--export-py-generated-sources` is enabled.

The codegen products are staged to a `__pants_codegen__` directory in
the exported digest, and then a helper Python script moves the codegen
products into the site-packages directory. A helper script was used
since it has access to the location of the site-packages directory
programmatically via the `site` module.