add workspace environment support (#20900)
Add support for a "workspace environment" which is similar to a local
environment except that execution happens in the repository itself
(i.e., the "workspace") and not in an execution sandbox. This implements
the user-visible experience of Design B from the ["In-Workspace Process
Execution" design
document](https://docs.google.com/document/d/1jUZTQHmUBr-Ij0nXOO0QX2Bq6XNANJg4-GNzhiOL4Xs/edit?usp=sharing).

This PR builds on top of the Rust support for in-workspace execution
contained in https://github.com/pantsbuild/pants/pull/20772.

Users will use the new `workspace_environment` target type to configure
workspace execution support.

The [pants-workspace-exec-testing
repository](https://github.com/tdyas/pants-workspace-exec-testing)
contains an example of how to use this support to allow Pants to invoke
Bazel in the workspace.