Upgrade default `pyright` version from 1.1.316 to 1.1.365 (#21007)
Pyright is rather out-of-date at the current version, so this PR updates
the default to the latest release as of 1 week ago.

In addition to the simple change made in `subsystem.py` I've also
included some changes to the tests. The example lockfile was updated
based on the results of running `npm i --save-dev pyright@1.1.365` in an
empty directory using Node 16 and `npm` 8.19.4.