fix regression in choosing a compatible local environment (Cherry pick of #21011) (#21014)
Introduction of workspace environments also introduced a regression in
the rule which selects a local environment compatible with the current
platform: If you have both a local environment and a workspace
environment defined in your configuration, then Pants will error that it
is ambiguous which one to select as a _local_ environment. This is wrong
of course; the rule should have ignored the workspace environment for
that logic.

The issue is that the rule was looking for targets with the
`CompatiblePlatformsField` which is now true of both `local_environment`
and `workspace_environment` targets.

Solve the issue by subclassing `CompatiblePlatformsField` for
`local_environment` target type and have the rule check for that. This
is what was done for `workspace_environment` as well so fits nicely with
that solution.