scala source file should not depend on scalac plugins (#20971)
These are compile-time-only dependencies and should not be packaged with
the source file or included in the classpath for the source file.

In our repository, this caused a conflict because the semanticdb package
internally wraps the fastparse library with a different version than the
one we are using. There is no need for semanticdb to be packaged or run
within a test of a target; it is a compile-time-only dependency.
Moreover, the scalac compile goal handles the scalac plugins on its own
and doesn't use that.

I ran this on our Scala repository, and it compiled successfully.