Add "workspace invalidation" sources support for shell / adhoc backends (#21051)
Add support for "workspace invalidation" sources for the `adhoc_tool`
and `shell_command` target types. This supports allows those targets to
depend on the content of files in the repository without materializing
those sources in the execution sandbox. This support is intended to be
used in conjunction with the workspace environment where execution does
not take place in a sandbox.

The new field `workspace_invalidation_sources` on both target types is a
list of globs into the repository. The digest of the referenced files
will be inserted as an environment variable in the process executed
(which makes it part of the process's cache key).