Sketch new "Maintenance tasks and scripts" contributing dev docs (#21061)
Fixes #21034 

This adds a new page to the "contributing > development" section of the
docs, that describes various common tasks that happen as part of the
development flow. This is partly reminding us of various scripts in
`build-support/bin/`, but not entirely.

This is just a starting point: it doesn't list all the scripts in
`build-support/bin/` (although some aren't intended for human use
anyway, I believe), probably doesn't list all the tasks. Easy to expand
in future!

Rendered table of contents: 


![image](https://github.com/pantsbuild/pants/assets/1203825/99b5b0b4-a820-4a47-95ac-473270e7471c)