add intrinsic rule to request metadata about paths in filesystem (#20996)
Add a new `path_metadata_request` intrinsic rule to allow rule code to
request metadata about paths in the filesystem.

The intended uses are to support:
- Some form of metadata-based invalidation for `adhoc_tool` and
`shell_command` targets, as explored in the proof of concept in
https://github.com/pantsbuild/pants/pull/20914. (This PR in fact is
split out from that other PR.)
- Future potential work to avoid indefinite negative caching of
PATH-style lookups (which is a problem of the existing BinaryPath
lookups) by switching to direct monitoring of system-level paths. This
API would likely be extended to that use case. (But I am not yet
committing to any particular solution. Still, some form of metadata API
will be useful.)

Metadata is represented by the `PathMetadata` dataclass.
`PathMetadataRequest` and `PathMetadataResult` are the input/output
types, respectively, for the intrinsic.

Note: `NodeKey::fs_path_to_watch` is introduced to allow
`NodeKey::PathMetadata` to watch the parent directory even though
`fs_subject` remains its configured path. This is necessary because the
watch code will error with "file not found" if the path to be watched
does not exist. The solution is to watch the parent directory and wait
for creation / removal events related to the `fs_subject`.