Always capture a `corepack` installation and support omitting version strings (#21021)
When a user _just_ provided 
```toml
[nodejs]
package_manager = 'pnpm'
```
and not providing which version to use, `pants` would invoke `corepack`
with illegal arguments. The intention was to support this mode, and
defer the decision of what version of `pnpm` the user gets to
`corepack`, as it curates releases of package managers.
This is now achieved by passing the `--all` flag for these cases.

Additionally, when the package manager is configured via a
`package.json` file not located at the build root, `corepack` outputs
would not be captured, resulting in failures because corepack would
attempt to download the package managers into what was by then an
immutable digest.
This is resolved by removing the `working_directory` and input digest
logic from the invocation. Since versions are parsed prior, there is
never a case that the `package.json` is read by `corepack` where pants
hasn't already parsed its contents.


Fixes https://github.com/pantsbuild/pants/issues/21020

---------

Co-authored-by: Huon Wilson <huon@exoflare.io>