Remove //BUILD_ROOT dependency from RuleRunner
Repos that use pants.testutil might not have a BUILD_ROOT
file, and most likely would not make any RuleRunner-based tests
depend on that file (adding it to the pytest sandbox). Instead,
the RuleRunner needs to manage adding a sentinel file.

This was not an issue before NativeOptionParser was wired up
in the python layer as part of 2.22. Now, however, NativeOptionParser
needs to be able to detect the build root using either an env var
(PANTS_BUILDROOT_OVERRIDE) or a sentinel file in the cwd.
This makes RuleRunner add that sentinel file.