Upgrade default Node.js version from 16.x to 18.x series (#21012)
Related Slack discussion here:
https://pantsbuild.slack.com/archives/C0D7TNJHL/p1717531611290729

This turned out to be pretty gnarly, especially when it came to
`corepack` + regenerating lock files. Most of the diff here is just
regenerating lockfiles based on the new package manager versions below.

Versions of the different package managers this was tested with:

* `npm`: `v10.8.1`
* `pnpm`: `8.15.8` (8.x)
* `yarn`: `1.22.2` (1.x)

### Process to make this PR

1. Bump the default Node.js version in `nodejs.py` along with the
`default_known_versions`.
2. Replace various tool lockfiles to work with the versions of the
package managers shown above.
3. Ensuring that a `packageManager` field was added into `package.json`
for use with Node and `corepack`.
4. Trying to mess with the lockfiles and regenerate until things work
properly.