Javascript backend - run node package install once per workspace (#20826)
Fixes #19236 and fixes #19235.

I've been poking around at this backend for a while trying to understand
it. I think I'm most of the way there...
I wanted to raise this PR now for feedback - I think it's fairly
complete as is - but I'm not sure if I have any blind spots in both
understanding the original design intent and how it's being used.

There is a lot more functionality than I expected going in - huge props
to @tobni for that - it would be great to fix the remaining issues and
get some docs out so it can be used more & we can get more feedback.

## Current process

For each first party package, the backend runs 'npm install' on its
first party dependency packages. This is a recursive process.

Each `npm install` command is executed in the context of that package,
generating a 'node_modules' directory containing 1st and 3rd party deps
(in separate directories). Source code (from `javascript_sources`
targets) is then copied in.
The output of `node_modules` from any 1st party dependency installs is
merged to form `node_modules` for the package.

### Problems

Per #19235, newer versions of package managers generate integrity files
in each node_modules directories. These cause conflicts on merge.

Secondly, and more critically, it seems (I've not dug too much into
this, but some past discussion
[here](https://pantsbuild.slack.com/archives/C0D7TNJHL/p1685026807993119)),
that each `npm install` is effectively installing all transient deps.

E.g. in the case of a tree like package-A -> package-B -> package-C ->
package-D ... I think while the intent is for package-A install to only
install direct dependents, it's installing all transitive, so there's a
lot more IO than intended. Effectively, it is merging ((package-A +
trans deps) + (package-B + trans deps) + (package-c + trans deps)...))

## New approach

Perform one install for the package in the workspace context. 
The install command is `npm install --workspace package-a`. 

The install command needs to have visibility of the root package.json,
which contains the 'workspaces' configuration, as well as any dependent
package.json. Currently this is all package.json in the project.

The output is a single node_modules directory at the project level
(assuming default package manager config - hoisting options could be
configured differently).

### Advantages

Simpler and is similar to how the user would use the package manager
natively to install in a mono-repo setup.
The backend is simpler to troubleshoot as there is only one sandbox per
install.
We delegate more of the package resolution process to the package
managers.
Performance should be theoretically better given the issue above - but
I've not tested. If anyone has anything handy for testing, please let me
know.

### Disadvantages

We are not caching each module install within Pants, which I'm guessing
was the original intent. But I'm not sure we're gaining much from this
approach currently.

---------

Co-authored-by: Rhys Madigan <rhys.madigan@sourse.co>