Native dockerfile parser (#21160)
Adds a dockerfile parser that aims to be ~compliant with the pex-python
one, but avoids installing the latter tooling.

1. I found that `tree-sitter-dockerfile` did not support value-less
options to instructions, so I've opened a fix for this.
2. I focused on being 1:1 with the old parser to the extent that
tree-sitter could (see
https://github.com/pantsbuild/pants/pull/21160#discussion_r1674400959)
without falling back to just parsing strings. Except for ARG, because
the python variant kept quotation I opted to mirror this in rust and the
tree-sitter traversal was not useful.

2 means that we allocate, duplicate and interpolate strings
unnecessarily often. I'd like to create a cleaner more rust-y data
structure in the future, but I think keeping the parser outputs similar
is more important right now.

---------

Co-authored-by: Huon Wilson <wilson.huon@gmail.com>