use a template for computing `output_path` for packageable targets (#21175)
As described in https://github.com/pantsbuild/pants/issues/18659, when a
user sets `output_path` on a packageable target, the user gives up the
default behavior which uses a normalized version of the target's
directory as the directory into which the place the built artifact.
While the user could replicate the behavior manually, it would be a
better user experience if the user can still have partial access to some
of the default behavior.

This PR teaches the `output_path` field how to derive its value from a
template instead of hard-coding its behavior in Python. The
`output_path` field learns the following template replacements:

- `spec_path_normalized`: The path to the target's directory ("spec
path") with forward slashes replaced by dots.
- `target_name_normalized`: The target's name 
- `file_suffix`: For target's which produce single file artifacts, this
is the file type suffix to use. This is empty if not applicable.

The default value for the `output_path` field is now
`"${spec_path_normalized}/${target_name_normalized}${file_suffix}"`
which replicates the existing behavior.

`string.Template` is used as the template engine.

This PR also fixes some incorrect usages of `output_path` in several
backends.